backend/requirements.txt
FastAPI
SQLAlchemy
pymysql
cryptography
pydantic
uvicorn
python-dotenv
requests